fix bugs

+++ libraries/aircraft-simulator/src/tower.cpp
@@ -25,7 +25,7 @@ double Tower::calculateTime(const OurVector<3> &current_position)
     const OurVector<3>& b = _position;
     const OurVector<3>& a = current_position;
 
-    double distance = sqrtl((std::pow((a[0] - b[0]), 2)) + (std::pow((a[1] - b[1]), 2)) +
+    double distance = sqrt((std::pow((a[0] - b[0]), 2)) + (std::pow((a[1] - b[1]), 2)) +
                                (std::pow((a[2] - b[2]), 2)));
 
     return distance / k_wave_speed;
